fixed creation of custom attributes
