+++ client/ayon_ftrack/lib/__init__.py
@@ -7,14 +7,6 @@ from .constants import (
     CUST_ATTR_INTENT,
     FPS_KEYS
 )
-from .custom_attributes import (
-    default_custom_attributes_definition,
-    app_definitions_from_app_manager,
-    tool_definitions_from_app_manager,
-    get_openpype_attr,
-    query_custom_attributes
-)
-
 from . import credentials
 from .utils import statics_icon
 
@@ -27,13 +19,6 @@ __all__ = (
     "CUST_ATTR_APPLICATIONS",
     "CUST_ATTR_INTENT",
     "FPS_KEYS",
-
-    "default_custom_attributes_definition",
-    "app_definitions_from_app_manager",
-    "tool_definitions_from_app_manager",
-    "get_openpype_attr",
-    "query_custom_attributes",
-
 
     "credentials",
