+++ ftrack_common/custom_attributes.py
@@ -1,3 +1,5 @@
+import os
+import json
 import itertools
 import collections
 
@@ -164,3 +166,61 @@ def get_custom_attributes_by_entity_id(
             entity_values[attr_id] = value
 
     return output
+
+
+def default_custom_attributes_definition():
+    """Default custom attribute definitions created in ftracl.
+
+    Todos:
+        Convert to list of dictionaries to be able determine order. Check if
+            ftrack api support to define order first!
+
+    Returns:
+        dict[str, Any]: Custom attribute configurations per entity type that
+            can be used to create/update custom attributes.
+    """
+
+    json_file_path = os.path.join(
+        os.path.dirname(os.path.abspath(__file__)),
+        "custom_attributes.json"
+    )
+    with open(json_file_path, "r") as json_stream:
+        data = json.load(json_stream)
+    return data
+
+
+# TODO applications and tools should be moved to 'applications' addon
+def app_definitions_from_app_manager(app_manager):
+    _app_definitions = []
+    for app_name, app in app_manager.applications.items():
+        if app.enabled:
+            _app_definitions.append(
+                (app_name, app.full_label)
+            )
+
+    # Sort items by label
+    app_definitions = []
+    for key, label in sorted(_app_definitions, key=lambda item: item[1]):
+        app_definitions.append({key: label})
+
+    if not app_definitions:
+        app_definitions.append({"empty": "< Empty >"})
+    return app_definitions
+
+
+def tool_definitions_from_app_manager(app_manager):
+    _tools_data = []
+    for tool_name, tool in app_manager.tools.items():
+        _tools_data.append(
+            (tool_name, tool.label)
+        )
+
+    # Sort items by label
+    tools_data = []
+    for key, label in sorted(_tools_data, key=lambda item: item[1]):
+        tools_data.append({key: label})
+
+    # Make sure there is at least one item
+    if not tools_data:
+        tools_data.append({"empty": "< Empty >"})
+    return tools_data
